Move classifiers to different directory
- create __init__ file with __all__ variable which indicate which classifiers
will be imported with 'from classifiers import *' import
- change imports (our local imports should be separately from imports from other packages)
