+++ sequence-classification/classifiers/__init__.py
@@ -1,0 +1,7 @@
+from .hmm_classifier import HMMClassifier
+from .knn_classifier import KNNClassifier
+from .neural_networks_classifier import NeuralNetworksClassifier
+from .svm_classifier import SVMClassifier
+
+
+__all__ = ['HMMClassifier', 'KNNClassifier', 'SVMClassifier', 'NeuralNetworksClassifier']
