+++ sequence-classification/classifiers/hmm_classifier.py
@@ -1,26 +1,26 @@
-from sequence_classifier import SequenceClassifier
-from sequence_transformer import SequenceTransformer
+from seqlearn.hmm import MultinomialHMM
 from sklearn.feature_extraction.text import CountVectorizer
 
-from sklearn.svm import LinearSVC
+from .sequence_classifier import SequenceClassifier
+from sequence_transformer import SequenceTransformer
 
 
-class SVMClassifier(SequenceClassifier):
-    def __init__(self, name='SVM'):
-        super(SVMClassifier, self).__init__(name)
-        self.model = LinearSVC(random_state=0, tol=1e-5)
+class HMMClassifier(SequenceClassifier):
+    def __init__(self, name='HMM'):
+        super(HMMClassifier, self).__init__(name)
+        self.model = MultinomialHMM()
 
     def fit(self, X_train, y_train):
-        self.model.fit(X_train, y_train)
+        self.model.fit(X_train, y_train, len(X_train))
 
     def predict(self, X):
         return self.model.predict(X)
 
     def get_transformer(self):
-        return SVMTransformer()
+        return HMMTransformer()
 
 
-class SVMTransformer(SequenceTransformer):
+class HMMTransformer(SequenceTransformer):
     def __init__(self):
         self.vectorizer = CountVectorizer()
