+++ sequence-classification/classifiers/svm_classifier.py
@@ -1,25 +1,26 @@
-from sequence_classifier import SequenceClassifier
-from sequence_transformer import SequenceTransformer
+from sklearn.svm import LinearSVC
 from sklearn.feature_extraction.text import CountVectorizer
 
-from seqlearn.hmm import MultinomialHMM
+from .sequence_classifier import SequenceClassifier
+from sequence_transformer import SequenceTransformer
 
-class HMMClassifier(SequenceClassifier):
-    def __init__(self, name='HMM'):
-        super(HMMClassifier, self).__init__(name)
-        self.model = MultinomialHMM()
+
+class SVMClassifier(SequenceClassifier):
+    def __init__(self, name='SVM'):
+        super(SVMClassifier, self).__init__(name)
+        self.model = LinearSVC(random_state=0, tol=1e-5)
 
     def fit(self, X_train, y_train):
-        self.model.fit(X_train, y_train, len(X_train))
+        self.model.fit(X_train, y_train)
 
     def predict(self, X):
         return self.model.predict(X)
 
     def get_transformer(self):
-        return HMMTransformer()
+        return SVMTransformer()
 
 
-class HMMTransformer(SequenceTransformer):
+class SVMTransformer(SequenceTransformer):
     def __init__(self):
         self.vectorizer = CountVectorizer()
